+++ CHANGELOG.md
@@ -2,3 +2,27 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com) and this project adheres to [Semantic Versioning](https://semver.org).
+
+## 2.0.0 - 2021-02-18
+
+### Added
+
+- Nothing
+
+### Changed
+
+- curl request
+- sendSms() added $curlOptions as parameter so that you can pass your own cURL options
+
+### Deprecated
+
+- Nothing
+
+### Removed
+
+- params key in response
+- remove CURLOPT_NOBODY, CURLOPT_HEADER in default cURL options
+
+### Fixed
+
+- Nothing